+++ app/models/PhpbbUser.php
@@ -1,24 +1,23 @@
 <?php
-class Phpbb extends Eloquent {
+class PhpbbUser extends Eloquent {
     
     public $connection = 'phpbb';
     public $timestamps = false;
     protected $guarded = array();
-    protected $table = 'phpbb_users';
 
     public static function changePassword($id, $email, $password)
     {
-        $user = Phpbb::where('user_email', $email)->first();
+        $user = PhpbbUser::where('user_email', $email)->first();
         if ($user) {
             if ($user->user_password != md5($password))
-                Phpbb::where('user_email', $email)->update(array('user_password' => md5($password)));
+                PhpbbUser::where('user_email', $email)->update(array('user_password' => md5($password)));
         }
         else
         {
             $lu = User::find($id);
             $lui = UserInfo::find($id);
-            $user = new Phpbb;
-            $user->user_id = Phpbb::max('user_id') + 1;
+            $user = new PhpbbUser;
+            $user->user_id = PhpbbUser::max('user_id') + 1;
             $user->user_active = 0;
             $user->username = $lui->name . " " . $lui->surname;
             $user->user_password = md5($password);
@@ -31,7 +30,7 @@ class Phpbb extends Eloquent {
     {
         $lu = User::find($id);
         $lui = UserInfo::find($id);
-        if (Phpbb::where('user_email', $lu->email)->update(array('user_active' => 1)))
+        if (PhpbbUser::where('user_email', $lu->email)->update(array('user_active' => 1)))
             return 1;
         else
             return 0;