+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## master
 - Improve file structure
+- Move the Model Phpbb to PhpbbUser
 
 ## v0.1.0
 - First version